moved memorypak functions to seperate module
